requirements.txt
tensorflow_hub
tensorflow-gpu